move remaining dependency tests
