+++ cargo-anatomy/tests/cli.rs
@@ -520,120 +520,3 @@ fn init_creates_config() {
     assert!(contents.contains("abstract_min"));
 }
 
-#[test]
-fn path_dep_import_dependency() {
-    let dir = tempfile::tempdir().unwrap();
-    std::fs::create_dir_all(dir.path().join("a/src")).unwrap();
-    std::fs::create_dir_all(dir.path().join("b/c/src")).unwrap();
-    std::fs::create_dir_all(dir.path().join("b/src")).unwrap();
-
-    std::fs::write(
-        dir.path().join("a/Cargo.toml"),
-        "[package]\nname = \"a\"\nversion = \"0.1.0\"\nedition = \"2024\"\n",
-    )
-    .unwrap();
-    std::fs::write(dir.path().join("a/src/lib.rs"), "pub mod foo;\n").unwrap();
-    std::fs::write(
-        dir.path().join("a/src/foo.rs"),
-        "#[derive(Debug)]\npub enum Foo { A }\n",
-    )
-    .unwrap();
-
-    std::fs::write(
-        dir.path().join("b/Cargo.toml"),
-        "[package]\nname = \"b\"\nversion = \"0.1.0\"\nedition = \"2024\"\n\n[dependencies]\na = { path = \"../a\" }\nc = { path = \"./c\" }\n",
-    )
-    .unwrap();
-    std::fs::write(dir.path().join("b/src/lib.rs"), "pub use c;\n").unwrap();
-
-    std::fs::write(
-        dir.path().join("b/c/Cargo.toml"),
-        "[package]\nname = \"c\"\nversion = \"0.1.0\"\nedition = \"2024\"\n\n[dependencies]\na = { path = \"../../a\" }\n",
-    )
-    .unwrap();
-    std::fs::write(
-        dir.path().join("b/c/src/lib.rs"),
-        "use a::foo;\npub const X: foo::Foo = foo::Foo::A;\n",
-    )
-    .unwrap();
-
-    std::fs::write(
-        dir.path().join("Cargo.toml"),
-        "[workspace]\nmembers = [\"a\", \"b\"]\nresolver = \"3\"\n",
-    )
-    .unwrap();
-
-    let mut cmd = Command::cargo_bin("cargo-anatomy").unwrap();
-    cmd.arg("-a").current_dir(dir.path());
-    let out = cmd.assert().get_output().stdout.clone();
-    let v: serde_json::Value = serde_json::from_slice(&out).unwrap();
-    let arr = v.get("crates").unwrap().as_array().unwrap();
-    let mut map = std::collections::HashMap::new();
-    for item in arr {
-        let pkg = item["crate_name"].as_str().unwrap();
-        map.insert(pkg, item);
-    }
-    assert_eq!(map["c"]["metrics"]["ce"].as_u64().unwrap(), 1);
-    assert_eq!(map["a"]["metrics"]["ca"].as_u64().unwrap(), 1);
-}
-
-#[test]
-fn external_crate_alias() {
-    let dir = tempfile::tempdir().unwrap();
-    let external = tempfile::tempdir().unwrap();
-    std::fs::create_dir_all(external.path().join("src")).unwrap();
-    std::fs::write(
-        external.path().join("Cargo.toml"),
-        "[package]\nname = \"foo_bar\"\nversion = \"0.1.0\"\n",
-    )
-    .unwrap();
-    std::fs::write(external.path().join("src/lib.rs"), "pub struct FooBar;\n").unwrap();
-
-    std::fs::create_dir(dir.path().join("app")).unwrap();
-    std::fs::create_dir(dir.path().join("app/src")).unwrap();
-    std::fs::write(
-        dir.path().join("app/Cargo.toml"),
-        format!(
-            "[package]\nname = \"app\"\nversion = \"0.1.0\"\n[dependencies]\nfoo_bar = {{ path = \"{}\" }}\n",
-            external.path().display()
-        ),
-    )
-    .unwrap();
-    std::fs::write(
-        dir.path().join("app/src/lib.rs"),
-        "use foo_bar as foo; pub struct App { d: foo::FooBar }\n",
-    )
-    .unwrap();
-
-    std::fs::write(
-        dir.path().join("Cargo.toml"),
-        "[workspace]\nmembers = [\"app\"]\n",
-    )
-    .unwrap();
-
-    let mut cmd = Command::cargo_bin("cargo-anatomy").unwrap();
-    cmd.args(["-a", "-x"]).current_dir(dir.path());
-    let out = cmd.assert().get_output().stdout.clone();
-    let v: serde_json::Value = serde_json::from_slice(&out).unwrap();
-    let obj = v.as_object().unwrap();
-    assert!(obj["meta"]["cargo-anatomy"]["target"].is_string());
-    let arr = obj.get("crates").unwrap().as_array().unwrap();
-    let mut map = std::collections::HashMap::new();
-    for item in arr {
-        let pkg = item["crate_name"].as_str().unwrap();
-        map.insert(pkg, item);
-    }
-    let app = map.get("app").unwrap();
-    let dep = map.get("foo_bar").unwrap();
-    assert_eq!(app["metrics"]["ce"].as_u64().unwrap(), 1);
-    assert_eq!(dep["metrics"]["ca"].as_u64().unwrap(), 1);
-    assert_eq!(dep["metrics"]["ce"].as_u64().unwrap(), 0);
-    assert_eq!(
-        app["details"]["external_depends_on"]["App"]["foo_bar"]
-            .as_array()
-            .unwrap()[0]
-            .as_str()
-            .unwrap(),
-        "FooBar"
-    );
-}
